Update jest, ts-jest and typescript versions
The versions of jest, ts-jest, and typescript in package.json have been updated for maintaining up-to-date dependencies and ensures compatibility with new features or bug fixes from these packages. The exact versions are now being used instead of the version range to avoid the potential issues that might be introduced by minor updates.